Update collision to improve performance by reducing get-ecs-component calls
Previously, procedure to get components (physic-2d and point-2d (by
calc-global-point)) is done in double loop for collision. But because
getting component is not so light, it largely affects to
peformance. So this commits improves it by getting them before double
loop (single loop).

For reference, this changes performance as the following.

*** Collision process only time; About 45-50 * 2 entities exist ***

- Before: about  33ms per frame
- After:  about 2.5ms per frame